+++ src/lib/definition.ts
@@ -5,13 +5,27 @@ interface PLSQLRoot {
 
 interface PLSQLSymbol {
     name: string;
+    definition?: string;
+    documentation?: string;
+    formatedDoc?: {
+        isMarkdown: boolean,
+        text: string
+    };
     offset?: number;
     offsetEnd?: number;
     kind: PLSQLSymbolKind;
     kindName: string;
+    params?: PLSQLParam[];
     symbols?: PLSQLSymbol[];
     parent?: PLSQLSymbol;
     root?: PLSQLRoot;
+}
+
+interface PLSQLParam {
+    text: string;
+    name?: string;
+    type: string;
+    kind: PLSQLParamKind;
 }
 
 const enum PLSQLSymbolKind {
@@ -27,4 +41,12 @@ const enum PLSQLSymbolKind {
     subtype,
     cursor,
     exception
+}
+
+const enum PLSQLParamKind {
+    none = 0,
+    return,
+    in,
+    out,
+    inout
 }