+++ src/lib/regEx/RegExParser.ts
@@ -9,16 +9,22 @@ export default class RegExpParser {
     public static regExpB: RegExp;
     public static regExpJumpEnd: RegExp;
     public static regExpJumpAsIs: RegExp;
+    public static regExpJumpDoc: RegExp;
+    public static regExpParams: RegExp;
 
     private static regComment = `(?:\\/\\*[\\s\\S]*?\\*\\/)|(?:--.*)`;
+    private static regCommentDoc = `(?:\\/\\*(\\*)?[\\s\\S]*?\\*\\/)|(?:--.*)`;
     private static regCommentInside = `(?:\\/\\*[\\s\\S]*?\\*\\/\\s*)?`; // a bit slower !
+    private static regJumpDoc = `(\\/\\*\\*[\\s\\S]*?\\*\\/)`;
 
+    private static REG_WORD = "[\\w\\$#]";
     private static regSymbolsCreate = `(?:(create)(?:\\s+or\\s+replace)?\\s+)?`;
     private static regSymbols = `(?:\\b(function|procedure|package)\\b(?:\\s+(body))?)\\s+`;
-    private static regSymbolsName = `(?:\"?\\w+\"?\\.)?\"?(\\w+)\"?`;
+    private static regSymbolsName = `(?:\"?${RegExpParser.REG_WORD}+\"?\\.)?\"?(${RegExpParser.REG_WORD}+)\"?`;
 
-    private static regSpecSymbols = `(?:(\\w+)\\s+(\\w+)\\s*(?:\\s*;|.[^;]*;))`;
-    private static regBody = `(?:\\b(procedure|function)\\b\\s+(\\w+)[\\s\\S]*?(;|\\b(?:is|as|begin)\\b))`;
+    private static regSpecSymbols = `(?:(${RegExpParser.REG_WORD}+)\\s+(${RegExpParser.REG_WORD}+)\\s*(?:\\s*;|.[^;]*;))`;
+    private static regBody = `(?:\\b(procedure|function)\\b\\s+(${RegExpParser.REG_WORD}+)[\\s\\S]*?(;|\\b(?:is|as|begin)\\b))`;
+    private static regParams = `(?:\\(|,)\\s*((${RegExpParser.REG_WORD}+)\\s*(in\\s+out|in|out)?\\s*(${RegExpParser.REG_WORD}*))|(?:\\breturn\\b\\s*(${RegExpParser.REG_WORD}*))`;
 
     private static regJumpEnd = `(\\bbegin|case\\b)|(?:(\\bend\\b)\\s*(?:\\b(if|loop|case)\\b)?)`;
     private static regJumpAsIs = `\\b(is|as)\\b`;
@@ -31,10 +37,12 @@ export default class RegExpParser {
             `${RegExpParser.regSymbolsName})`;
 
         this.regExp = new RegExp(regExpParser, 'gi');
-        this.regExpS = new RegExp(`${this.regComment}|${`(\\b(?:end|create)\\b)`}|${this.regSpecSymbols}`, 'gi');
-        this.regExpB = new RegExp(`${this.regComment}|${this.regBody}|(\\bbegin\\b)|${this.regSpecSymbols}`, 'gi');
+        this.regExpS = new RegExp(`${this.regCommentDoc}|${`(\\b(?:end|create)\\b)`}|${this.regSpecSymbols}`, 'gi');
+        this.regExpB = new RegExp(`${this.regCommentDoc}|${this.regBody}|(\\bbegin\\b)|${this.regSpecSymbols}`, 'gi');
         this.regExpJumpEnd = new RegExp(`${this.regComment}|${this.regJumpEnd}`, 'gi');
         this.regExpJumpAsIs = new RegExp(`${this.regJumpAsIs}`, 'gi');
+        this.regExpJumpDoc = new RegExp(`${this.regJumpDoc}\\s*$`, 'gi');
+        this.regExpParams = new RegExp(`${this.regParams}`, 'gi');
     }
 
     public static getSymbols(text: string, fileName?: string): PLSQLRoot  {
@@ -80,6 +88,7 @@ export default class RegExpParser {
                     offset = this.regExp.lastIndex;
                     fromOffset = this.jumpAsIs(text, offset);
                     if (fromOffset !== offset) { // if equal => no is|as => continue on same level...
+                        symbol.definition = found[0] + text.substring(offset, fromOffset-2);
                         offset = fromOffset;
                         if (symbol.kind === PLSQLSymbolKind.packageSpec)
                             offset = this.getSymbolsSpec(text, offset, symbol);
@@ -95,6 +104,31 @@ export default class RegExpParser {
         }
 
         return root;
+    }
+
+    public static parseParams(symbol: PLSQLSymbol): PLSQLParam[] {
+        if (symbol.params || !symbol.definition)
+            return symbol.params;
+
+        symbol.params = [];
+
+        let found;
+        while (found = this.regExpParams.exec(symbol.definition)) {
+            if (found[1])
+                symbol.params.push({
+                    text: found[1],
+                    name: found[2],
+                    type: found[4],
+                    kind: this.getParamKind(found[3]) // todo: convert
+                });
+            else // if (found[5])
+                symbol.params.push({
+                    text: found[0],
+                    type: found[5],
+                    kind: PLSQLParamKind.return
+                });
+        }
+        return symbol.params;
     }
 
     /// (only for spec)
@@ -117,16 +151,27 @@ export default class RegExpParser {
     private static getSymbolsSpec(text: string, fromOffset: number, parent: PLSQLSymbol): number  {
         let found,
             lastIndex = fromOffset,
+            lastDoc = null,
             symbol;
 
         this.regExpS.lastIndex = lastIndex;
         while (found = this.regExpS.exec(text)) {
-            if (found[2] && found[3]) {
-                symbol = this.createSymbolItem(found[2], found[3], found.index, parent, false);
-                symbol.offsetEnd = this.regExpS.lastIndex;
-            } else if (found[1]) // end || create
+            if (found[1]) { // doc
+                lastDoc = found.index;
+                continue;
+            }
+            else if (found[3] && found[4]) {
+                symbol = this.createSymbolItem(found[3], found[4], found.index, parent, false);
+                if (symbol) {
+                    symbol.definition = found[0];
+                    symbol.offsetEnd = this.regExpS.lastIndex;
+                    if (lastDoc != null)
+                        symbol.documentation = this.jumpDoc(text, lastDoc, found.index);
+                }
+            } else if (found[2]) // end || create
                 break;
             lastIndex = this.regExpS.lastIndex;
+            lastDoc = null;
         }
         return lastIndex;
     }
@@ -134,46 +179,58 @@ export default class RegExpParser {
     private static getSymbolsBody(text: string, fromOffset: number, parent: PLSQLSymbol, extractSymbol = true): number  {
         let found, symbol: PLSQLSymbol,
             lastIndex = fromOffset,
+            lastDoc = null,
             oldIndex, isBody;
 
         this.regExpB.lastIndex = lastIndex;
         while (found = this.regExpB.exec(text)) {
+            if (found[1]) { // doc
+                lastDoc = found.index;
+                continue;
+            }
             oldIndex = lastIndex;
             lastIndex = this.regExpB.lastIndex;
-            if (found[4]) // begin
+            if (found[5]) // begin
                 break;
-            else if (found[5] && found[6]) {
+            else if (found[6] && found[7]) {
                 if (extractSymbol) {
-                    symbol = this.createSymbolItem(found[5], found[6], found.index, parent, false);
-                    if (symbol)
+                    symbol = this.createSymbolItem(found[6], found[7], found.index, parent, false);
+                    if (symbol) {
+                        symbol.definition = found[0];
                         symbol.offsetEnd = lastIndex;
-                    else {
+                        if (lastDoc != null)
+                            symbol.documentation = this.jumpDoc(text, lastDoc, found.index);
+                    } else {
                         // if it's not a symbol, something goes wrong => break
                         lastIndex = oldIndex;
                         break;
                     }
                 }
-            } else if (found[1] && found[2]) { // function, procedure
+            } else if (found[2] && found[3]) { // function, procedure
                 // Declare function, procedure => add symbol
                 if (!parent.symbols)
                     parent.symbols = [];
-                isBody = found[3].toLowerCase() !== ';';
-                symbol = this.createSymbolItem(found[1], found[2], found.index, parent, isBody);
-                if (isBody)
-                    if (found[3].toLowerCase() === 'begin') {
-                        // begin => jump to end
-                        lastIndex = this.jumpToEnd(text, lastIndex);
-                        symbol.offsetEnd = lastIndex;
-                        this.regExpB.lastIndex = lastIndex;
-                    } else { // is,as
-                        // read between is and begin (subPro/subFunc)
-                        lastIndex = this.getSymbolsBody(text, lastIndex, symbol, false);
+                isBody = found[4].toLowerCase() !== ';';
+                symbol = this.createSymbolItem(found[2], found[3], found.index, parent, isBody);
+                if (symbol) {
+                    symbol.definition = found[0];
+                    if (lastDoc != null)
+                        symbol.documentation = this.jumpDoc(text, lastDoc, found.index);
+                    if (isBody) {
+                        if (found[4].toLowerCase() === 'begin') {
+                            // begin => jump to end
+                        } else { // is,as
+                            // read between is and begin (subPro/subFunc)
+                            lastIndex = this.getSymbolsBody(text, lastIndex, symbol, false);
+                        }
                         // jump to end
                         lastIndex = this.jumpToEnd(text, lastIndex);
                         symbol.offsetEnd = lastIndex;
                         this.regExpB.lastIndex = lastIndex;
                     }
                 }
+            }
+            lastDoc = null;
         }
         return lastIndex;
     }
@@ -218,6 +275,18 @@ export default class RegExpParser {
         if (match)
             return this.regExpJumpAsIs.lastIndex;
         return fromOffset;
+    }
+
+    private static jumpDoc(text: string, fromOffset: number, toOffset: number): string {
+        // find doc above a symbol
+        let match;
+
+        text = text.substr(0, toOffset);
+        this.regExpJumpDoc.lastIndex = fromOffset;
+        match = this.regExpJumpDoc.exec(text);
+        if (match)
+            return match[0];
+        return '';
     }
 
     private static jumpToEnd(text: string, fromOffset: number): number {
@@ -273,5 +342,18 @@ export default class RegExpParser {
             return PLSQLSymbolKind.exception;
         else
             return PLSQLSymbolKind.variable;
+    }
+
+    private static getParamKind(type: string): PLSQLParamKind {
+        if (type === 'in')
+            return PLSQLParamKind.in;
+        if (type === 'out')
+            return PLSQLParamKind.out;
+        if (type === 'inout')
+            return PLSQLParamKind.inout;
+        if (type === 'return')
+            return PLSQLParamKind.return;
+        else
+            return PLSQLParamKind.none;
     }
 }