+++ test/sql/src/PACKAGES/PCK$1.sql
@@ -1,0 +1,5 @@
+CREATE OR REPLACE PACKAGE "schema"."PCK$1"
+as
+  procedure my$Procedure(param1 in varchar2)
+end;
+/